chore(integrations/intercom): extract handler in dedicated file (#14435)
